requirements.txt
streamlit>=1.24.0
pandas>=1.5.0
plotly>=5.10.0
numpy>=1.23.0
openpyxl>=3.0.0
gspread>=5.7.0
gspread-dataframe>=3.2.0
google-auth>=2.0.0
